+++ src/noahtheduke/splint/parser.clj
@@ -69,7 +69,7 @@
                    (if-let [resolved-ns (get-in @ns-state [:aliases ns-str])]
                      resolved-ns
                      (if (= :current ns-str)
-                       "splint-auto-current"
+                       (or (:current @ns-state) "splint-auto-current")
                        (str "splint-auto-alias-" (name ns-str)))))
    :postprocess (fn postprocess [{:keys [obj loc]}]
                   (when-let [{:keys [current aliases imports]} (parse-ns obj)]