+++ CHANGELOG.md
@@ -24,6 +24,10 @@ This changelog is loose. Versions are not semantic, they are incremental. Splint
 - Disable `lint/thread-macro-one-arg` by default. It harms readability in a lot of cases and has limited usefulness.
 - `:import` parsing now includes both the base class name as well as the fully qualified class name in the returned map, which improves all interop scenarios.
 
+### Fixed
+
+- Correctly resolve auto-resolving keywords in the current namespace: `(ns foo) ::bar` will be parsed as `:foo/bar` instead of `:splint-auto-current/bar`.
+
 ## 1.20.0 - 2025-03-28
 
 ### New Rules